refactor(app): move member removal to workspace repository
